New arguments and speed improvements
Added arguments to get the version of the script, update the installomator and patchomator scripts if they need updated, and to set a script proxy if necessary.

Changed the version check of the app labels to use 'is-at-least'.

Rearranged installomator check and added a fatal check if installomator is not installed and the script is being run in install mode.

Created some arrays for the labels to not require a second lookup if needed.

Added a script check to the verifyApp function. If the script is installed from GitHub it downloads the latest pkg or the version installed pkg and does an md5 file check and a spctl or pkgutil check for expectedTeamID.

Added a summary after the appVerify that will let the user know how many of the apps that were found need updates. Only available if not in quiet mode.

Miscellaneous cleanup mostly involving tabs and spaces.