Dialog Updates, Timeouts, and MDM
Added timeout for label replacement prompts and included a prompt via Swift Dialog. Default 120, can be set via --timeout (10-9000)

Added the --icon argument for setting the Swift Dialog icon.

Fixed the moveable argument for Swift Dialog to allow the command to be moved around.

Revamped the config file logic.

Added MDM option to set the Logos. More settings can be added in the future to that logic in the future.

Revamped the CLIOptions to allow for them to overwrite default options, and made all default options get pushed to Installomator even if CLIOptions exists.